// modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModelBuilder/ModelBuilderContext/typesEnum.ts
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export enum TYPES {
	ADD_OBJECT_DEFINITION_TO_OBJECT_FOLDER = 'ADD_OBJECT_DEFINITION_TO_OBJECT_FOLDER',
	ADD_OBJECT_FIELD = 'ADD_OBJECT_FIELD',
	BULK_CHANGE_NODE_VIEW = 'BULK_CHANGE_NODE_VIEW',
	CHANGE_MODAL_VISIBILITY = 'CHANGE_MODAL_VISIBILITY',
	CHANGE_NODE_VIEW = 'CHANGE_NODE_VIEW',
	DELETE_OBJECT_FIELD = 'DELETE_OBJECT_FIELD',
	SET_DELETE_OBJECT_DEFINITION = 'SET_DELETE_OBJECT_DEFINITION',
	SET_ELEMENTS = 'SET_ELEMENTS',
	SET_LOADING_OBJECT_FOLDER = 'SET_LOADING_OBJECT_FOLDER',
	SET_OBJECT_FOLDER_NAME = 'SET_OBJECT_FOLDER_NAME',
	SET_SELECTED_OBJECT_DEFINITION_NODE = 'SET_SELECTED_OBJECT_DEFINITION_NODE',
	SET_SELECTED_OBJECT_DEFINITION_NODE_POSITION = 'SET_SELECTED_OBJECT_DEFINITION_NODE_POSITION',
	SET_SELECTED_OBJECT_FIELD = 'SET_SELECTED_OBJECT_FIELD',
	SET_SELECTED_OBJECT_RELATIONSHIP_EDGE = 'SET_SELECTED_OBJECT_RELATIONSHIP_EDGE',
	SET_SHOW_CHANGES_SAVED = 'SET_SHOW_CHANGES_SAVED',
	SET_SHOW_SIDEBARS = 'SET_SHOW_SIDEBARS',
	UPDATE_MODEL_BUILDER_STRUCTURE = 'UPDATE_MODEL_BUILDER_STRUCTURE',
	UPDATE_OBJECT_DEFINITION_NODE = 'UPDATE_OBJECT_DEFINITION_NODE',
	UPDATE_OBJECT_FIELD_NODE_ROW = 'UPDATE_OBJECT_FIELD_NODE_ROW',
}